# pkg/fileutils: GetTotalUsedFds: reduce allocations